+++ src/main/java/io/company/brewcraft/repository/PurchaseOrderRepository.java
@@ -17,4 +17,4 @@ public interface PurchaseOrderRepository extends JpaRepository<PurchaseOrder, Lo
     @Override
     @Modifying
     @Query("delete from purchase_order p where p.id in (:ids)")
-    int deleteByIds(Iterable<Long> ids);}
+    long deleteByIds(Iterable<Long> ids);}
