+++ src/test/java/io/company/brewcraft/service/impl/InvoiceServiceTest.java
@@ -160,7 +160,7 @@ public class InvoiceServiceTest {
    public void testDelete_CallsRepoServiceDeleteBulk_WhenInvoiceExists() {
        doReturn(123).when(this.mRepoService).delete(Set.of(1L, 2L, 3L));
 
-       final int count = this.service.delete(Set.of(1L, 2L, 3L));
+       final long count = this.service.delete(Set.of(1L, 2L, 3L));
        assertEquals(123, count);
    }
