+++ src/test/java/io/company/brewcraft/service/impl/PurchaseOrderServiceTest.java
@@ -139,7 +139,7 @@ public class PurchaseOrderServiceTest {
     public void testDelete_CallsRepoServiceDeleteBulk_WhenPurchaseOrderExists() {
         doReturn(123).when(this.mRepoService).delete(Set.of(1L, 2L, 3L));
 
-        final int count = this.service.delete(Set.of(1L, 2L, 3L));
+        final long count = this.service.delete(Set.of(1L, 2L, 3L));
         assertEquals(123, count);
     }
