+++ src/test/java/io/company/brewcraft/service/impl/ShipmentServiceTest.java
@@ -169,7 +169,7 @@ public class ShipmentServiceTest {
    public void testDelete_CallsRepoServiceDeleteBulk_WhenShipmentExists() {
        doReturn(123).when(this.mRepoService).delete(Set.of(1L, 2L, 3L));
 
-       final int count = this.service.delete(Set.of(1L, 2L, 3L));
+       final long count = this.service.delete(Set.of(1L, 2L, 3L));
        assertEquals(123, count);
    }
